wip: first analysis work
